requirements.txt

ipython
jupyter
attrs
Cython
gensim
click
